added user roles, permissions

+++ src/main/java/com/example/demo/BookController.java
@@ -1,22 +1,26 @@
 package com.example.demo;
 
+import org.apache.coyote.BadRequestException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
+import java.util.*;
 
 @RestController
 @RequestMapping("/books")
 public class BookController {
-
     private final BookService bookService;
+    private final UserRepository userRepository;
+    private final UserDetailsServiceImpl userDetailsService;
 
     @Autowired
-    public BookController(BookService bookService) {
+    public BookController(BookService bookService, UserRepository userRepository) {
         this.bookService = bookService;
+        this.userRepository = userRepository;
+        this.userDetailsService = new UserDetailsServiceImpl(userRepository);
     }
 
     @GetMapping
@@ -31,7 +35,12 @@ public class BookController {
         String username = authentication.getName();
         System.out.println("User " + username + " is attempting to add a book.");
         try {
-            return ResponseEntity.ok(bookService.saveBook(book));
+            boolean hasPermission = userDetailsService.hasPermission(username, Collections.singletonList("ADMIN"));
+            if (hasPermission) {
+                return ResponseEntity.ok(bookService.saveBook(book));
+            } else {
+                throw new BadRequestException("User not allowed to add book");
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).build(); // Return 400 if an exception occurs
@@ -52,7 +61,12 @@ public class BookController {
         String username = authentication.getName();
         System.out.println("User " + username + " is attempting to update book with id: " + id);
         try {
-            return ResponseEntity.ok(bookService.updateBook(id, bookDetails));
+            boolean hasPermission = userDetailsService.hasPermission(username, Collections.singletonList("ADMIN"));
+            if (hasPermission) {
+                return ResponseEntity.ok(bookService.updateBook(id, bookDetails));
+            } else {
+                throw new BadRequestException("User not allowed to update book");
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).build(); // Return 400 if an exception occurs
@@ -64,34 +78,47 @@ public class BookController {
         String username = authentication.getName();
         System.out.println("User " + username + " is attempting to delete book with id: " + id);
         try {
-            bookService.deleteBook(id);
-            return ResponseEntity.noContent().build();
+            boolean hasPermission = userDetailsService.hasPermission(username, Collections.singletonList("ADMIN"));
+            if (hasPermission) {
+                bookService.deleteBook(id);
+                return ResponseEntity.noContent().build();
+            } else {
+                throw new BadRequestException("User not allowed to delete book");
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).build(); // Return 400 if an exception occurs
         }
     }
 
-    // New endpoint for renting a book
     @PutMapping("/rent/{id}")
     public ResponseEntity<Book> rentBook(@PathVariable String id, Authentication authentication) {
         String username = authentication.getName();
         System.out.println("User " + username + " is attempting to rent book with id: " + id);
         try {
-            return ResponseEntity.ok(bookService.rentBook(id));
+            boolean hasPermission = userDetailsService.hasPermission(username, Arrays.asList("USER", "ADMIN"));
+            if (hasPermission) {
+                return ResponseEntity.ok(bookService.rentBook(id));
+            } else {
+                throw new BadRequestException("User not allowed to rent book");
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null); // Return 400 if an exception occurs
         }
     }
 
-    // New endpoint for returning a book
     @PutMapping("/return/{id}")
     public ResponseEntity<Book> returnBook(@PathVariable String id, Authentication authentication) {
         String username = authentication.getName();
         System.out.println("User " + username + " is attempting to return book with id: " + id);
         try {
-            return ResponseEntity.ok(bookService.returnBook(id));
+            boolean hasPermission = userDetailsService.hasPermission(username, Arrays.asList("USER", "ADMIN"));
+            if (hasPermission) {
+                return ResponseEntity.ok(bookService.returnBook(id));
+            } else {
+                throw new BadRequestException("User not allowed to return book");
+            }
         } catch (Exception e) {
             e.printStackTrace();
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null); // Return 400 if an exception occurs
@@ -99,10 +126,15 @@ public class BookController {
     }
 
     @GetMapping("/rented")
-    public ResponseEntity<List<Book>> getRentedBooks(Authentication authentication) {
+    public ResponseEntity<List<Book>> getRentedBooks(Authentication authentication) throws Exception {
         String username = authentication.getName();
         System.out.println("User " + username + " is fetching all rented books.");
-        List<Book> rentedBooks = bookService.getAllRentedBooks();
-        return ResponseEntity.ok(rentedBooks);
+        boolean hasPermission = userDetailsService.hasPermission(username, Arrays.asList("USER", "ADMIN"));
+        if (hasPermission) {
+            List<Book> rentedBooks = bookService.getAllRentedBooks();
+            return ResponseEntity.ok(rentedBooks);
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 }
